Remove the basic_response file
